% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{fmpc_13f_cik_name}
\alias{fmpc_13f_cik_name}
\title{Find name of company by cik}
\usage{
fmpc_13f_cik_name(cik = "0001067983")
}
\arguments{
\item{cik}{a valid cik}
}
\value{
company name associated with cik
}
\description{
Ciks are unique IDs for each company that files a 13F. This function returns
the name of the company associated with one or more ciks
}
\examples{
\dontrun{
# Function can work without a valid API token
fmpc_set_token() # defaults to 'demo'
fmpc_13f_cik_name(cik = '0001067983')
}
}
