% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{fmpc_cots_symbols}
\alias{fmpc_cots_symbols}
\title{Get list of COTS symbols}
\usage{
fmpc_cots_symbols()
}
\value{
a data frame of cots symbols with descriptions
}
\description{
Provides a list of all the symbols available on the Commitment of Traders
data pull
}
\examples{

\dontrun{

# Must set a valid API token
fmpc_set_token('FMPAPIKEY')
fmpc_cots_symbols()

}
}
