% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{fmpc_economic_events}
\alias{fmpc_economic_events}
\title{List economic events that can be searched}
\usage{
fmpc_economic_events()
}
\value{
economic metrics
}
\description{
Shows the economic events that can be fed into
\code{\link{fmpc_economic_results}}
}
\examples{
\dontrun{
# Function can work without a valid API token
 fmpc_set_token('FMPAPIKEY')
fmpc_economic_events()
}
}
