% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{fmpc_financial_zip}
\alias{fmpc_financial_zip}
\title{Get a list of Financial ZIP file locations}
\usage{
fmpc_financial_zip()
}
\value{
a data frame of symbols with zip file location
}
\description{
Get location to download ZIP file for all symbols. This function will not
download the file. Use a 'wget' command or enter the link into a browser.
}
\examples{

\dontrun{

# Must set a valid API token
fmpc_set_token('FMPAPIKEY')
fmpc_financial_zip()

}
}
