% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secDetails.R
\name{fmpc_sec_filings}
\alias{fmpc_sec_filings}
\title{SEC filings}
\usage{
fmpc_sec_filings(symbols = c("AAPL"), type = NULL, limit = 500)
}
\arguments{
\item{symbols}{one or more symbols. Use NULL for general latest news}

\item{type}{filing type, e.g. '10-K' (optional)}

\item{limit}{limit results for each ticker}
}
\value{
data frame of ratings data
}
\description{
SEC filings
}
