% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityMetrics.R
\name{fmpc_security_sector}
\alias{fmpc_security_sector}
\title{Pull daily returns for all sectors}
\usage{
fmpc_security_sector(days = 25)
}
\arguments{
\item{days}{number of trading days to pull returns from current day. For
example, 20 will pull the last 20 trading days, about one month of data.}
}
\value{
data frame of sectors and daily returns
}
\description{
For each day, the return for a sector will be pulled
}
\examples{
\dontrun{

fmpc_set_token('FMPAPIKEY')
allSymbs = fmpc_security_sector(30)

}
}
