% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketInfo.R
\name{fmpc_symbol_search}
\alias{fmpc_symbol_search}
\title{Search for symbols or companies}
\usage{
fmpc_symbol_search(query = "apple inc", limit = 10)
}
\arguments{
\item{query}{a search string,for example 'tech'}

\item{limit}{limit the results to a certain number}
}
\value{
list of symbols and companies that meet the search criteria
}
\description{
Enter search string to pull back matching symbols or companies
}
\examples{

\dontrun{
# Function can work without a valid API token
fmpc_set_token() # defaults to 'demo'
fmpc_symbol_search('apple')
fmpc_symbol_search('tech', 100)
}
}
