% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketInfo.R
\name{fmpc_symbols_available}
\alias{fmpc_symbols_available}
\title{Get symbols available through FMP Cloud that have a profile}
\usage{
fmpc_symbols_available()
}
\value{
a data frame of symbols with descriptions
}
\description{
Use a valid API Token to pull all symbols that have a profile on FMP Cloud.
Pull the profile using \code{\link{fmpc_security_profile}}. Set API
token using \code{\link{fmpc_set_token}}.
}
\examples{

\dontrun{

# Must set a valid API token
fmpc_set_token('FMPAPIKEY')
fmpc_symbols_available()

}
}
