% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketInfo.R
\name{fmpc_symbols_sp500}
\alias{fmpc_symbols_sp500}
\title{Current or historical S&P 500 constituents}
\usage{
fmpc_symbols_sp500(period = c("current", "historical"))
}
\arguments{
\item{period}{'current' for current list, 'historical' for a list of
companies that have been added and the ones that were replaced}
}
\value{
data frame of S&P 500 constituents
}
\description{
Shows current or historical companies in the S&P 500
}
\examples{

\dontrun{

# Must set a valid API token
fmpc_set_token('FMPAPIKEY')
fmpc_symbols_sp500()
fmpc_symbols_sp500('historical')
}

}
